Fix
- แก้ไขแสดงไม่ขึ้น Get popular (day)
- single announce image full path
- get announce category filter

+++ src/controllers/announcement/update.ts
@@ -95,8 +95,8 @@ const updateAnnounce = async (req: any, res: Response) => {
       data: error.details,
     });
   }
-
-  if(update != null && oldImage){
+  let checkimgdel = req.file?.filename
+  if(update != null && oldImage && checkimgdel != null && checkimgdel != "" && checkimgdel != undefined ){
     unlink(`././src/image/image-Announcement/${oldAnnouncement.Image}`, () => {});
   }
